refactor: move block consolidation logic to validations
