+++ validations/Cargo.toml
@@ -11,6 +11,7 @@ failure = "0.1.8"
 itertools = "0.8.2"
 log = "0.4.8"
 
+witnet_config = { path = "../config" }
 witnet_crypto = { path = "../crypto" }
 witnet_data_structures = { path = "../data_structures" }
 witnet_rad = { path = "../rad" }
